+++ src/aranet_cloud/__init__.py
@@ -54,7 +54,7 @@ from .models import (
     Unit,
 )
 
-__version__ = "0.2.1"
+__version__ = "0.2.2"
 
 __all__ = [
     "Alarm",